back-scrapper/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asttokens==3.0.1
async-timeout==4.0.3
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
coreapi==2.3.3
coreschema==0.0.4
decorator==5.2.1
distro==1.9.0
Django==5.2.7
django-cors-headers==4.9.0
django-redis==6.0.0
django-rest-swagger==2.2.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
drf-yasg==1.21.11
exceptiongroup==1.3.0
executing==2.2.1
frozenlist==1.8.0
genai==2.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.55.0
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
inflection==0.5.1
ipython==8.37.0
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-core==0.3.79
langchain-text-splitters==0.3.11
langsmith==0.4.34
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
multidict==6.7.0
openai==0.27.10
openapi-codec==1.3.2
orjson==3.11.3
packaging==25.0
parso==0.8.5
pexpect==4.9.0
playwright==1.55.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.0
pydantic_core==2.41.1
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.1.3
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.28.0
rsa==4.9.1
simplejson==3.20.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.3.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
zstandard==0.25.0
